finished build V1.0
All functionality has been implemented. The Moderator Panel is completely independent from the rest of the system due to the IModeratorPanel (or smt) interface. If you want to make a connection to a remote database (for cluster for instance) then you should implement an instance of the IModeratorPanel and attach it to the main menu, during startup